Moved text to readme, lowercased piStatus files
