+++ src/app.js
@@ -74,14 +74,6 @@ db((connection) => {
     console.log(`Started on port ${app.server.address().address}:${app.server.address().port}`);
   });
 
-  // Requires running mosquitto broker server on localhost:1883
-  // for OS X
-  //
-  // brew install mosquitto
-  //
-  // for ubuntu
-  //
-  // sudo apt-get install mosquitto
   mqttListener();
 });
