+++ src/models/index.js
@@ -1,5 +1,5 @@
 export {default as Lanemap} from './lanemap';
 export {default as User} from './user';
 export {default as Penalty} from './penalty';
-export {default as piStatus} from './piStatus';
+export {default as piStatus} from './pistatus';
 export {default as WhitePlate} from './whiteplate';
